requirements.txt
requests~=2.32.3
notion-client~=2.3.0
python-dotenv~=1.0.1